Better Naming and Architecture
