+++ src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Project from './App';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
 import CssBaseline from '@material-ui/core/CssBaseline';
 
 ReactDOM.render(
   <React.Fragment>
     <CssBaseline />
-    <Project />
+    <App />
   </React.Fragment>,
   document.getElementById('app-container')
 );
